pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "mud-balance-ui"
version = "0.1.0"
description = "Web UI for the headless balance simulator of bylins/mud (issue #2967)"
requires-python = ">=3.11"
dependencies = [
    "Flask>=3.0",
    "ruamel.yaml>=0.18",
    "ulid-py>=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]
prod = [
    "gunicorn>=22.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "B", "UP", "RUF"]
ignore = ["E501"]
